requirements.txt
pytube
rich
typer
typing_extensions
ffmpeg-python
